Disable Tor when user cancel Tor bootstrap
When user cancel Tor bootstrap, it is reasonable to assume user wants to disable Tor. For example, when users misconfigure the Tor, Tor bootstrap may stuck at a point. After user hit the cancel button, we should not let Tor keep trying to connect to the Tor network with the misconfigured settings.